requirements.txt
pytest
pytest-playwright
playwright
allure-pytest
Faker